turn Models and LLM into pydantic-compatible classes
required as they're part of the config class
